requirements.txt
Flask
requests
numpy
joblib
keras
tensorflow-cpu
pillow
lightgbm
scikit-learn
seaborn


